vectors like float4, int4,..., etc. supported.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2612 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUCuda++/Testprog/stereoTest/main.cc
@@ -72,8 +72,8 @@ main(int argc, char* argv[])
     using namespace	TU;
 
     using pixel_type	= u_char;
-    using score_type	= short;
-  //using score_type	= float;
+  //using score_type	= short;
+    using score_type	= float;
     
     std::string	paramFile		= DEFAULT_PARAM_FILE;
     std::string	configDirs		= DEFAULT_CONFIG_DIRS;
